+++ src/main/java/lotto/LottoMachine.java
@@ -1,5 +1,12 @@
 package lotto;
 
+import lotto.domain.LottoCount;
+import lotto.domain.LottoMoney;
+import lotto.domain.LottoResults;
+import lotto.domain.LottoTicket;
+import lotto.domain.LottoTickets;
+import lotto.exception.LottoException;
+import lotto.factory.LottoTicketFactory;
 import view.InputView;
 import view.OutputMessage;
 import view.ResultView;
