+++ src/main/java/lotto/exception/LottoErrorCode.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.exception;
 
 import view.Printable;
