+++ src/main/java/lotto/exception/LottoException.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.exception;
 
 public class LottoException extends RuntimeException {
     public LottoException(LottoErrorCode errorCode) {
