+++ src/test/java/lotto/exception/LottoErrorCodeTest.java
@@ -1,7 +1,6 @@
-package lotto;
+package lotto.exception;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
